refactoring to cubit auth
