+++ lib/domain/services/auth_data_storage.dart
@@ -1,0 +1,4 @@
+class AuthDataStorage {
+  String login = '';
+  String password = '';
+}
