+++ lib/ui/widgets/auth/auth_widget.dart
@@ -3,7 +3,7 @@
 // import 'package:comics_db_app/app_colors.dart';
 // import 'package:comics_db_app/library/widgets/inherited/notifier_provider.dart';
 // import 'package:comics_db_app/resources/resources.dart';
-// import 'package:comics_db_app/ui/widgets/auth/auth_model.dart';
+// import 'package:comics_db_app/ui/widgets/auth/auth_view_cubit.dart';
 // import 'package:flutter/material.dart';
 //
 // class AuthWidget extends StatefulWidget {
