+++ lib/ui/widgets/loader_widget/loader_view_cubit.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-
 import 'package:bloc/bloc.dart';
 import 'package:comics_db_app/domain/blocs/auth_bloc.dart';
 
@@ -10,10 +9,15 @@ class LoaderViewCubit extends Cubit<LoaderViewCubitState> {
   late final StreamSubscription<AuthState> authBlocSubscription;
 
   LoaderViewCubit(LoaderViewCubitState initialState, this.authBloc) : super(initialState) {
-    authBloc.add(AuthCheckStatusEvent());
+    // if (authBloc.state is AuthAuthorizedState) {
+    //   emit(LoaderViewCubitState.authorized);
+    // } else {
     _onState(authBloc.state);
     authBlocSubscription = authBloc.stream.listen(_onState);
+    authBloc.add(AuthCheckStatusEvent());
   }
+
+  // }
 
   void _onState(AuthState state) {
     if (state is AuthAuthorizedState) {
